[rbp/omxplayer] Apply deinterlace changes immediately
Currently on Pi the deinterlace state is only sampled when opening the player.
This means to change deinterlace settings, you need to change in the video OSD then manually quit and restart the video.
This commit detects a change in current deinterlace settings and closes and opens the video player which applies the setting.

For seekable streams, seeking to "now" makes the video restart quicker.
Without the seek, the video does reappear, but takes a number of seconds