[rbp/omxplayer] Fixes for trickplay
In #2955 we lied in GetTime when in trickplay modes to avoid jittery updates in elapsed time.
This was wrong when GetClock() didn't match the media time (e.g. when playback was started from a bookmark).

The jitter is caused by the DISPLAYTIME messages that don't really work on omxplayer.
Instead bypass the DISPLAYTIME messages and set the time explicity in OMXPlayer (using media time where appropriate).

Also in trickplay, do the seek/flush step more to fix an issue when going from normal to rewind after EOS

Also don't honour the "accurate" seek flag, as it results in corrupt frames being displayed