[omx] Skip the resize when not needed when decoding jpegs
The decode to texture path almost always uses cached jpegs that are the correct size, so the resize is rarely needed.
The re-enc path usually needs resizing, but may not where the source is small.

Skipping the resize stage saves a little time and memory on GPU and also saves some setup time.