[omxcodec] Fix for stills with hardware codec
Hardware codecs typically run asynchronously to the decoded frames.
The reason that hardware codecs don't work when hint.stills is true
is because DVDPlayerVideo stops calling Decode/GetPicture when there
are no more decoded frames.

This means the still frames never get read out.
The solution is to also call Decode/GetPicture with the timeout messages
when hint.stills is true.