[omxplayer] Don't flush queued data on general reset
We were flushing the queue of data to gpu (which can be a number of seconds worth) on a general reset.
However this is generated on inoccuous events like a chapter change in dvd playback.

This results in a few seconds loss of audio, and a temporary audio/video sync error