[mmalrenderer] Avoid submitting a buffer to renderer twice
Bad things happen when submitting a buffer to renderer twice. We were attempting to protect against that with
a flip count. However that was getting reset from AddProcessor, the the sequence:
Decode->AddProcessor->RenderUpdate->AddProcessor->RenderUpdate
could occur. Instead use a bit from the MMAL flags which is unlikely to get out of sync with the MMAL buffer.