[rbp] Disable analogue output of sync when passthrough is enabled
The analogue and output of Pi Sink doesn't support passthrough.
This also applies to the 'Both' (HDMI and Analogue) setting.
If passthrough is enabled in GUI you get black screen and errors:
https://discourse.osmc.tv/t/some-streams-are-staying-black/2311

We need to either disable passthrough or disable analogue.

I think the best solution to this is to force HDMI output when passthrough is active.

This allows a user to use receiver for videos with dts/ac3 and TV for music or lower quality formats.