[mmalcodec] Remove the decoder queue and use gpu timestamps
There are too many streams that don't have a 1:1 mapping of encoded video frames to decoded pictures.
If that isn't true, the dts queues gets out of sync with the real frames.

Interlaces streams and streams with errors cause issues.

Instead remove the dts queue and use the gpu's timestamps.