[omxplayer] Flush EOS message from queue to avoid it turning up after a seek
A GENERAL_EOS is queued when demuxer reaches EOF but playback may continue for several seconds.
It should be possible to seek backwards during this playback time, but the GENERAL_EOS may still be queued after flushing
and it unwantedly causes stream to stall after the seek. So flush these obsolete messages when flushing data packets