[PR 10695] [peripherals] Fix grabbing joystick if there's no add-on to translate input
This was causing a problem on the Shield where the controller would work
until it was scanned as a joystick. Without being scanned, the controller
sends keyboard presses, which avoid the need to be translated by an add-on.
Scanning the controller with no add-on present caused input to break, so
check this case before initializing the joystick.