Fix TextureCache::CacheImage does not fill texture outparam
If the texture out parameter is given, it's expected to be
filled.
However when requesting a texture which is already in the
processlist, the method fails to fill the requested texture.

This resulted in a hard to trigger bug, where the default
texture is rendered instead of the real thumbnail, although
the thumbnail is available.